requirements.txt
discord.py
jira
python-dotenv
pytz